% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_token.R
\name{dbh_api_token}
\alias{dbh_api_token}
\title{Returns token}
\usage{
dbh_api_token(brukernavn = "", passord = "")
}
\arguments{
\item{brukernavn}{username}

\item{passord}{password}
}
\value{
empty string for not token users or
content of \code{\link{dbh_api_token_get_new}}
}
\description{
Returns current token using global variable , or retrieves new token if expired
}
