% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_token.R
\name{dbh_api_token_get_new}
\alias{dbh_api_token_get_new}
\title{Retrieving new token}
\usage{
dbh_api_token_get_new(brukernavn, passord)
}
\arguments{
\item{brukernavn}{username}

\item{passord}{password}
}
\value{
return new DBH-APIs token
}
\description{
Retrieving new token
}
