% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_cache.R
\name{dbh_cache}
\alias{dbh_cache}
\title{Cache DBH tables}
\usage{
dbh_cache(
  table_id,
  filters = NULL,
  group_by = NULL,
  sort_by = NULL,
  exclude = NULL,
  variables = NULL,
  api_version = 1,
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  compress_file = TRUE
)
}
\arguments{
\item{table_id}{The code name for the dataset}

\item{filters}{A named list, where the names are the names of the variables
in the dataset to be filtered, and the values contain the matching
conditions.}

\item{group_by}{A list of variables to include in the aggregation for
aggregating tables.}

\item{sort_by}{A list of variables that define the sorting order.}

\item{exclude}{A named list, where the names must also occur in
\code{filters}, and the values specify values to be excluded from the
filter.}

\item{variables}{A list of variables to include in dataset}

\item{api_version}{DBH-API version}

\item{cache}{a logical value whether to cache}

\item{update_cache}{a logical value whether to update cache ,
default is without updating}

\item{cache_dir}{a path to cache directory.
Default value create a cache directory}

\item{compress_file}{logical value whether
to compress, default is to compress}
}
\value{
a tibble
}
\description{
Cache DBH tables
}
