% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_cache.R
\name{dbh_clean_cache}
\alias{dbh_clean_cache}
\title{Clean DBH cache}
\usage{
dbh_clean_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{a path to cache directory.}
}
\value{
{No return value}
}
\description{
Delete all .rds files from the DBH cache directory.
}
