% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{dbh_data}
\alias{dbh_data}
\title{Retrieve data from DBH-API as an R data frame}
\usage{
dbh_data(
  table_id,
  filters = NULL,
  group_by = NULL,
  sort_by = NULL,
  exclude = NULL,
  variables = NULL,
  api_version = 1
)
}
\arguments{
\item{table_id}{Numeric. Required. The unique ID of the dataset to query.}

\item{filters}{Optional. A named list where names correspond to dataset variables, and values specify filter criteria.}

\item{group_by}{Optional. A character vector of variables to use for aggregating tables.}

\item{sort_by}{Optional. A character vector of variables to define sorting order.}

\item{exclude}{Optional. A named list specifying values to exclude from the filters.}

\item{variables}{Optional. A character vector of variables to include when retrieving non-aggregated data.}

\item{api_version}{Integer. Version of the DBH-API to use (default is 1).}
}
\value{
A `data.frame` containing the requested dataset, with variables properly typed.
}
\description{
Sends a request to the DBH-API and returns the result as an R `data.frame`.
Data types are automatically converted based on the table metadata,
using the helper function \code{\link{dbh_metadata}}.

Authentication with the DBH-API is optional:
- If no token is provided, the API still returns data, but numeric values will be rounded (typically to the nearest 3).
- If a valid token is provided, full and precise data values are returned.

To authenticate, users must set their SSO ID and SSO Secret as environment variables
(\code{dbhapi_sso_id} and \code{dbhapi_sso_secret}) manually or in an `.Renviron` file.
If these environment variables are set, the package will automatically retrieve and use a token when sending requests.
}
\examples{
\dontrun{
# Example 1: Query with filters and group_by
filter_example <- dbh_data(
  table_id = 142,
  filters = list(Institusjonskode = c("top", 3)),
  group_by = "Institusjonskode"
)

# Example 2: Download bulk table
bulk_example <- dbh_data(table_id = 211)
}

}
\seealso{
\code{\link{dbh_metadata}}
}
