% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{dbh_metadata}
\alias{dbh_metadata}
\title{Get metadata for DBH datasets}
\usage{
dbh_metadata(table_id = NULL)
}
\arguments{
\item{table_id}{Numeric. Required. The unique ID of the dataset.}
}
\value{
A tibble containing variable metadata.
}
\description{
Gets information on variables
included in DBH datasets (type of variable, data type).
}
\examples{
\dontrun{
# Show metadata for table 88
meta_table <- dbh_metadata(88)
}
}
