% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{dbh_tabell}
\alias{dbh_tabell}
\title{Get data from dbh api as R dataframe}
\usage{
dbh_tabell(tabell_id, filters = NULL, group_by = NULL,
  sort_by = NULL, exclude = NULL, api_versjon = 1,
  statuslinje = "N", decimal_separator = readr::locale()$decimal_mark,
  meta = FALSE)
}
\arguments{
\item{tabell_id}{a code name for dataset}

\item{filters}{is the same as filters in dbh api: item, all, top, between, lessthan}

\item{group_by}{group by variables in the same way as in dbh api}

\item{sort_by}{sort variables}

\item{exclude}{varible values that we did not want to inculde in filtering}

\item{api_versjon}{defined api constat value 1}

\item{statuslinje}{defined api constatn value N}

\item{decimal_separator}{defined api value}

\item{meta}{is set to FALSE and does not return metadata}
}
\value{
R dataframe
}
\description{
@description
 A function send request from R to api and get data from api into R.
 Data are converted in right format using helepfunction dbh_metadata
 For token users it is possible to get token using function get_dbh_token and us it further
}
\examples{
 kandidatar<-dbh_tabell(907)
}
