% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{dbh_toc}
\alias{dbh_toc}
\title{Table of contents for DBH-API}
\usage{
dbh_toc(table_id = NULL)
}
\arguments{
\item{table_id}{A vector of code names for the datasets retrieve information for, or NULL to get the whole table of contents}
}
\value{
a tibble
}
\description{
Table of contents for DBH-API
}
