% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.dbh_content}
\alias{.dbh_content}
\title{Get content data from DBH-API}
\usage{
.dbh_content(table_id)
}
\arguments{
\item{table_id}{A vector of code names for the datasets to get variable information for}
}
\value{
A tibble
}
\description{
Get content data from DBH-API
}
\keyword{internal}
