% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.dbh_groupBy}
\alias{.dbh_groupBy}
\title{Title}
\usage{
.dbh_groupBy(table_id)
}
\arguments{
\item{table_id}{A vector of code names for the datasets to get variable information for}
}
\value{
A list of group by variables
}
\description{
Title
}
