% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.default_group_by}
\alias{.default_group_by}
\title{Default group_by for a table as suggested by DBH}
\usage{
.default_group_by(table_id)
}
\arguments{
\item{table_id}{A vector of code names for the datasets to get variable information for}
}
\value{
A list of group by variables

a list with group_by values
}
\description{
Default group_by for a table as suggested by DBH
}
\keyword{internal}
