% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.fetch_data}
\alias{.fetch_data}
\title{Generic function to fetch data from DBH API}
\usage{
.fetch_data(url)
}
\arguments{
\item{url}{URL to fetch the data from}
}
\value{
A tibble
}
\description{
Generic function to fetch data from DBH API
}
\keyword{internal}
