% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.get_metadata}
\alias{.get_metadata}
\title{Get table with metadata for variables in the DBH API}
\usage{
.get_metadata(table_id = NULL)
}
\description{
Get table with metadata for variables in the DBH API
}
\keyword{internal}
