% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.get_new_metadata}
\alias{.get_new_metadata}
\title{Download table with metadata for variables in DBH-API}
\usage{
.get_new_metadata()
}
\value{
a tibble
}
\description{
Download table with metadata for variables in DBH-API
}
\keyword{internal}
