% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.get_new_toc}
\alias{.get_new_toc}
\title{Download table of contents of the DBH API}
\usage{
.get_new_toc()
}
\value{
a tibble
}
\description{
Download table of contents of the DBH API
}
\keyword{internal}
