% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_token.R
\name{.get_new_token}
\alias{.get_new_token}
\title{Retrieve new JWT token for DBH-API}
\usage{
.get_new_token(sso_id, sso_secret)
}
\arguments{
\item{sso_id}{Client ID for DBH-API authentication}

\item{sso_secret}{Secret ID for DBH-API authentication}
}
\value{
A new JWT token for DBH-API
}
\description{
Retrieve new JWT token for DBH-API
}
\keyword{internal}
