% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{.get_toc}
\alias{.get_toc}
\title{Get table with content for variables in the DBH-API}
\usage{
.get_toc(table_id = NULL)
}
\description{
Get table with content for variables in the DBH-API
}
\keyword{internal}
