% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{.make_query}
\alias{.make_query}
\title{Create query for DBH-API}
\usage{
.make_query(table_id, filters, group_by, sort_by, exclude, variables)
}
\arguments{
\item{table_id}{The code name for the dataset}

\item{filters}{A named list, where the names are the names of the variables
in the dataset to be filtered, and the values contain the matching
conditions.}

\item{group_by}{A list of variables to include in the aggregation for
aggregating tables.}

\item{sort_by}{A list of variables that define the sorting order.}

\item{exclude}{A named list, where the names must also occur in
\code{filters}, and the values specify values to be excluded from the
filter.}

\item{variables}{A list of variables, do not combine with group_by}
}
\value{
A list of length containing the DBH-API query that can be converted
  to the proper JSON
}
\description{
Create queries to be converted to the DBH-API JSON format
}
\keyword{internal}
