% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_token.R
\name{dbh_api_token}
\alias{dbh_api_token}
\title{Returns current token using golobal variable variable, or retrieves new token if expired}
\usage{
dbh_api_token(brukernavn = "", passord = "")
}
\arguments{
\item{brukernavn}{username}

\item{passord}{password}
}
\value{
dbh api token
}
\description{
Returns current token using golobal variable variable, or retrieves new token if expired
}
