% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{dbh_json_query}
\alias{dbh_json_query}
\title{Send query to dbh api}
\usage{
dbh_json_query(tabell_id, filters = list(), group_by = list(),
  sort_by = list(), exclude = NULL)
}
\arguments{
\item{tabell_id}{a code name for dataset}

\item{filters}{is the same as filters in dbh api: item, all, top, between, lessthan}

\item{group_by}{group by variables in the same way as in dbh api}

\item{sort_by}{sort variables}

\item{exclude}{varible values that we did not want to inculde in filtering}
}
\value{
query
}
\description{
@description
Help function which send query form r to api, it is designes in the way that it looks like api query and allows
the same functionality as in api
}
