% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{dbh_metadata}
\alias{dbh_metadata}
\title{Get metadata for DBH datasets}
\usage{
dbh_metadata(table_id = NULL)
}
\arguments{
\item{table_id}{A vector of code names for the datasets to
get variable information for, or NULL to get info for all variables.}
}
\value{
A tibble
}
\description{
Gets information on variables
included in DBH datasets (type of variable, data type).
}
\examples{
meta_table <- dbh_metadata(88)
}
