% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{dbh_metadata}
\alias{dbh_metadata}
\title{Get api metadata for each table
 @description
A function to read into R metadata which are stored in api table 2}
\usage{
dbh_metadata(tabell_id)
}
\arguments{
\item{tabell_id}{a code name for dataset}
}
\value{
metadata as a table in csv format
}
\description{
Get api metadata for each table
 @description
A function to read into R metadata which are stored in api table 2
}
\examples{
dbh_metadata(88)

}
