% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_metadata.R
\name{dbh_metadata}
\alias{dbh_metadata}
\title{Get APIs metadata for each table}
\usage{
dbh_metadata(tabell_id)
}
\arguments{
\item{tabell_id}{a code name for dataset}
}
\value{
metadata for selected dataset as R dataframe
}
\description{
A function to read into R metadata  stored in APIs table 2
}
\examples{
dbh_metadata(88)
}
