% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{dbh_tabell}
\alias{dbh_tabell}
\title{Get data from API as R dataframe}
\usage{
dbh_tabell(
  tabell_id,
  filters = NULL,
  group_by = NULL,
  sort_by = NULL,
  exclude = NULL,
  api_versjon = 1,
  statuslinje = "N",
  decimal_separator = readr::locale()$decimal_mark,
  meta = FALSE
)
}
\arguments{
\item{tabell_id}{a code name for dataset}

\item{filters}{is the same as filters in DBH-API: item, all, top, between, greaterthan, lessthan}

\item{group_by}{group by variables in the same way as in DBH-API}

\item{sort_by}{sort variables in the same way as in DBH-API}

\item{exclude}{variable values we do not want to include in filtering}

\item{api_versjon}{defined DBH-API constant value 1}

\item{statuslinje}{defined DBH-API constant value N}

\item{decimal_separator}{defined DBH-API value}

\item{meta}{is set to FALSE and does not return metadata, set meta=TRUE and you will get metadata}
}
\value{
R dataframe
}
\description{
Send request from R to DBH-API and get data from DBH-API into R.
 Data are converted in right format using help function dbh_metadata \code{\link{dbh_metadata}}
 For token users it is possible to get token using function \code{\link{.get_token}} and use it further
}
\examples{
dbh_tabell(60, filters=list( "Institusjonskode"="1120","Alder"=c("between",c("30","40")))
, group_by="Alder")
}
