% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_api_tabell.R
\name{.make_query}
\alias{.make_query}
\title{Create a query object for the DBH-API}
\usage{
.make_query(
  table_id,
  filters = list(),
  group_by = list(),
  sort_by = list(),
  exclude = NULL,
  variables = list()
)
}
\arguments{
\item{table_id}{Numeric. The unique ID of the dataset to query.}

\item{filters}{Optional. A named list where each name corresponds to a variable to filter,
and each value specifies filter conditions.}

\item{group_by}{Optional. A character vector specifying variables to group by (used for aggregated tables).}

\item{sort_by}{Optional. A character vector specifying variables to sort the results by.}

\item{exclude}{Optional. A named list where each name corresponds to a variable,
and each value lists values to exclude from that variable's filter.}

\item{variables}{Optional. A character vector specifying variables to retrieve (used for non-aggregated tables).
Cannot be used together with \code{group_by}.}
}
\value{
A list representing the query structure ready to be converted to JSON format for the DBH-API.
}
\description{
Constructs a query list in the format required by the DBH-API,
which can then be serialized to JSON.
}
\keyword{internal}
